# MB-33455: improve ComputeGeoRange() performance