This change improves ComputeGeoRange() performance by avoiding trips
to the allocator by using pre-allocated []byte slices and by avoiding
creation of interim [][]byte slices by append()'ing directly to
returned output slices.

Before the change...
```
  BenchmarkComputeGeoRangePt01-8 100000     18516 ns/op      2005 B/op     63 allocs/op
  BenchmarkComputeGeoRangePt1-8   10000    108882 ns/op     96267 B/op    736 allocs/op
  BenchmarkComputeGeoRange10-8       50  35883331 ns/op  35812513 B/op 184543 allocs/op
  BenchmarkComputeGeoRange100-8      10 192568538 ns/op 187524856 B/op 926510 allocs/op
```

After the change...
```
  BenchmarkComputeGeoRangePt01-8 100000     12447 ns/op       280 B/op      7 allocs/op
  BenchmarkComputeGeoRangePt1-8   30000     47053 ns/op     16416 B/op     28 allocs/op
  BenchmarkComputeGeoRange10-8      100  11836988 ns/op   8503406 B/op     76 allocs/op
  BenchmarkComputeGeoRange100-8      20  72555603 ns/op  42778777 B/op     89 allocs/op
```

In some (perhaps unnatural) cases where the bounding boxes are very
large... performance improves by ~2x, memory utilization improves by
~4x and trips to the allocator improves by ~10000x.

See also: https://issues.couchbase.com/browse/MB-33455