@steveyen 2019-03-25
> I dislike this for all of the usual reasons.  It is a faster, but less obvious version of code I never understood in the first place when I did the initial port.

Same feeling -- in that I also don't fully understand 100% what this code is doing, but just did the same old playbook or "mechanical" optimization of the pre-allocated slices trick, etc.